Create blackjackprogramLeeTA.java
Program Request Derrick Dollesin:

Create a program where users are to play blackjack. 
A random number generator gives the player a card with a value of 1 - 11. 
The program then adds the two numbers and if they are above 21 the player loses immediately. 
If they are less the player is able to ask for another number but if over 21 to the sum then they lose. 
Do the same for a computer player who must ask for another number if sum of first two cards 
are less than 15. Player or computer hit 21 they win, or whoever is closer to 21 wins.

Author: Jonathan Lee
Professor: Gita Faroughi
Class: Sierra College CSCI13
Date: Feb 12 2022

This system works I would like options for multiple decks